requirements_venv310.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.2
kiwisolver==1.4.8
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.42.0
numpy==2.2.4
packaging==25.0
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
pyglet==2.1.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
six==1.17.0
sympy==1.14.0
tzdata==2025.2
